Fixed healthbar and made enemies attack properly
The enemies fist collider will be huge at the moment, until I figure out a good size for it.

+++ Assets/Scripts/Internal/Player/UI/LocalHealthbar.cs
@@ -16,7 +16,7 @@ public class LocalHealthbar : MonoBehaviour
 
     private void onTakeDamage(string[] args)
     {
-        healthbar.fillAmount = stats.health / stats.maxHealth; 
+        healthbar.fillAmount = (float)stats.health / stats.maxHealth; 
     }
 
     private void OnMenuChange(string[] args)
